# Support TLS protocol in kafka-producer-performance